Basic model / creation of summary table is done
Now we need to add style to the spreadsheet cells to make them easier to
read and have the summary table mirror the styling of pivot tables that
Excel creates

+++ src/net/mcnewfamily/rmcnew/shared/Util.java
@@ -73,7 +73,7 @@ public class Util {
                 if(DateUtil.isCellDateFormatted(cell)) {
                     value = cell.getDateCellValue().toString();
                 } else {
-                    value = Double.toString(cell.getNumericCellValue());
+                    value = Integer.toString((int)cell.getNumericCellValue());
                 }
                 break;
             case Cell.CELL_TYPE_BOOLEAN:
